Refactor git utilities
