+++ clowder/utility/git_utilities_private.py
@@ -1,0 +1,297 @@
+"""Git utilities"""
+import sys
+from git import Repo
+from termcolor import colored, cprint
+from clowder.utility.clowder_utilities import execute_command
+from clowder.utility.print_utilities import (
+    format_path,
+    format_ref_string,
+    format_remote_string,
+    print_command_failed_error,
+    print_error
+)
+
+# Disable errors shown by pylint for no specified exception types
+# pylint: disable=W0702
+# Disable errors shown by pylint for catching too general exception Exception
+# pylint: disable=W0703
+# Disable errors shown by pylint for too many statements
+# pylint: disable=R0915
+
+def _checkout_branch(repo_path, branch, remote, depth):
+    """Checkout branch, and create if it doesn't exist"""
+    repo = _repo(repo_path)
+    correct_branch = False
+    if branch in repo.heads:
+        default_branch = repo.heads[branch]
+        try:
+            not_detached = not repo.head.is_detached
+            same_branch = repo.head.ref == default_branch
+        except Exception as err:
+            pass
+        else:
+            if not_detached and same_branch:
+                print(' - On default branch')
+                correct_branch = True
+        finally:
+            if not correct_branch:
+                branch_output = format_ref_string(branch)
+                try:
+                    print(' - Checkout branch ' + branch_output)
+                    default_branch.checkout()
+                except Exception as err:
+                    message = colored(' - Failed to checkout branch ', 'red')
+                    print(message + branch_output)
+                    print_error(err)
+                    sys.exit(1)
+    else:
+        _create_local_tracking_branch(repo_path, branch, remote, depth)
+
+def _checkout_sha(repo_path, sha):
+    """Checkout commit by sha"""
+    repo = _repo(repo_path)
+    correct_commit = False
+    try:
+        same_sha = repo.head.commit.hexsha == sha
+        is_detached = repo.head.is_detached
+    except Exception as err:
+        pass
+    else:
+        if same_sha and is_detached:
+            print(' - On correct commit')
+            correct_commit = True
+    finally:
+        if not correct_commit:
+            commit_output = format_ref_string(sha)
+            try:
+                print(' - Checkout commit ' + commit_output)
+                repo.git.checkout(sha)
+            except Exception as err:
+                message = colored(' - Failed to checkout commit ', 'red')
+                print(message + commit_output)
+                print_error(err)
+                sys.exit(1)
+
+def _checkout_tag(repo_path, tag):
+    """Checkout commit tag is pointing to"""
+    repo = _repo(repo_path)
+    tag_output = format_ref_string(tag)
+    correct_commit = False
+    if tag in repo.tags:
+        try:
+            same_commit = repo.head.commit == repo.tags[tag].commit
+            is_detached = repo.head.is_detached
+        except Exception as err:
+            pass
+        else:
+            if same_commit and is_detached:
+                print(' - On correct commit for tag')
+                correct_commit = True
+        finally:
+            if not correct_commit:
+                try:
+                    print(' - Checkout tag ' + tag_output)
+                    repo.git.checkout(tag)
+                except Exception as err:
+                    message = colored(' - Failed to checkout tag ', 'red')
+                    print(message + tag_output)
+                    print_error(err)
+                    sys.exit(1)
+    else:
+        print(' - No existing tag ' + tag_output)
+
+def _create_checkout_branch(repo_path, branch, remote, depth):
+    """Create and checkout local branch"""
+    repo = _repo(repo_path)
+    remote_output = format_remote_string(remote)
+    print(' - Fetch from ' + remote_output)
+    if depth == 0:
+        command = ['git', 'fetch', remote, '--prune', '--tags']
+    else:
+        command = ['git', 'fetch', remote, '--depth', str(depth), '--prune', '--tags']
+    return_code = execute_command(command, repo_path)
+    if return_code != 0:
+        message = colored(' - Failed to fetch from ', 'red')
+        print(message + remote_output)
+        print_command_failed_error(command)
+        sys.exit(return_code)
+    branch_output = format_ref_string(branch)
+    try:
+        print(' - Create branch ' + branch_output)
+        default_branch = repo.create_head(branch)
+    except Exception as err:
+        message = colored(' - Failed to create branch ', 'red')
+        print(message + branch_output)
+        print_error(err)
+        sys.exit(1)
+    else:
+        try:
+            print(' - Checkout branch ' + branch_output)
+            default_branch.checkout()
+        except Exception as err:
+            message = colored(' - Failed to checkout branch ', 'red')
+            print(message + branch_output)
+            print_error(err)
+            sys.exit(1)
+
+def _create_local_tracking_branch(repo_path, branch, remote, depth):
+    """Create and checkout tracking branch"""
+    repo = _repo(repo_path)
+    branch_output = format_ref_string(branch)
+    remote_output = format_remote_string(remote)
+    try:
+        origin = repo.remotes[remote]
+    except Exception as err:
+        message = colored(' - No existing remote ', 'red')
+        print(message + remote_output)
+        print_error(err)
+        sys.exit(1)
+    else:
+        if depth == 0:
+            print(' - Fetch from ' + remote_output)
+            command = ['git', 'fetch', remote, '--prune', '--tags']
+        else:
+            print(' - Fetch from ' + remote_output + ' ' + branch_output)
+            command = ['git', 'fetch', remote, branch, '--depth', str(depth),
+                       '--prune', '--tags']
+        return_code = execute_command(command, repo_path)
+        if return_code != 0:
+            message = colored(' - Failed to fetch from ', 'red')
+            print(message + remote_output)
+            print_command_failed_error(command)
+            sys.exit(return_code)
+        try:
+            print(' - Create branch ' + branch_output)
+            default_branch = repo.create_head(branch, origin.refs[branch])
+        except Exception as err:
+            message = colored(' - Failed to create branch ', 'red')
+            print(message + branch_output)
+            print_error(err)
+            sys.exit(1)
+        else:
+            try:
+                print(' - Set tracking branch ' + branch_output +
+                      ' -> ' + remote_output + ' ' + branch_output)
+                default_branch.set_tracking_branch(origin.refs[branch])
+            except Exception as err:
+                message = colored(' - Failed to set tracking branch ', 'red')
+                print(message + branch_output)
+                print_error(err)
+                sys.exit(1)
+            else:
+                try:
+                    print(' - Checkout branch ' + branch_output)
+                    default_branch.checkout()
+                except Exception as err:
+                    message = colored(' - Failed to checkout branch ', 'red')
+                    print(message + branch_output)
+                    print_error(err)
+                    sys.exit(1)
+
+def _create_remote_tracking_branch(repo_path, branch, remote, depth):
+    """Create remote tracking branch"""
+    repo = _repo(repo_path)
+    branch_output = format_ref_string(branch)
+    remote_output = format_remote_string(remote)
+    try:
+        origin = repo.remotes[remote]
+    except Exception as err:
+        message = colored(' - No existing remote ', 'red')
+        print(message + remote_output)
+        print_error(err)
+        sys.exit(1)
+    else:
+        print(' - Fetch from ' + remote_output)
+        if depth == 0:
+            command = ['git', 'fetch', remote, '--prune', '--tags']
+        else:
+            command = ['git', 'fetch', remote, '--depth', str(depth),
+                       '--prune', '--tags']
+        return_code = execute_command(command, repo_path)
+        if return_code != 0:
+            message = colored(' - Failed to fetch from ', 'red')
+            print(message + remote_output)
+            print_command_failed_error(command)
+            sys.exit(return_code)
+        if branch in origin.refs:
+            try:
+                repo.git.config('--get', 'branch.' + branch + '.merge')
+            except:
+                message_1 = colored(' - Remote branch ', 'red')
+                message_2 = colored(' already exists', 'red')
+                print(message_1 + branch_output + message_2)
+                print()
+                sys.exit(1)
+            else:
+                print(' - Tracking branch ' + branch_output + ' already exists')
+        else:
+            try:
+                print(' - Push remote branch ' + branch_output)
+                repo.git.push(remote, branch)
+            except Exception as err:
+                message = colored(' - Failed to push remote branch ', 'red')
+                print(message + branch_output)
+                print_error(err)
+                sys.exit(1)
+            else:
+                try:
+                    print(' - Set tracking branch ' + branch_output +
+                          ' -> ' + remote_output + ' ' + branch_output)
+                    repo.active_branch.set_tracking_branch(origin.refs[branch])
+                except Exception as err:
+                    message = colored(' - Failed to set tracking branch ', 'red')
+                    print(message + branch_output)
+                    print_error(err)
+                    sys.exit(1)
+
+def _pull_remote_branch(repo_path, remote, branch):
+    """Pull from remote branch"""
+    repo = _repo(repo_path)
+    if not repo.head.is_detached:
+        try:
+            branch_output = format_ref_string(branch)
+            remote_output = format_remote_string(remote)
+            print(' - Pull latest changes from ' + remote_output + ' ' + branch_output)
+            print(repo.git.pull(remote, branch))
+        except Exception as err:
+            cprint(' - Failed to pull latest changes', 'red')
+            print_error(err)
+            sys.exit(1)
+
+def _ref_type(ref):
+    """Return branch, tag, sha, or unknown ref type"""
+    git_branch = "refs/heads/"
+    git_tag = "refs/tags/"
+    if ref.startswith(git_branch):
+        return 'branch'
+    elif ref.startswith(git_tag):
+        return 'tag'
+    elif len(ref) is 40:
+        return 'sha'
+    else:
+        return 'unknown'
+
+def _repo(repo_path):
+    """Create Repo instance for path"""
+    try:
+        repo = Repo(repo_path)
+    except Exception as err:
+        repo_path_output = format_path(repo_path)
+        message = colored("Failed to create Repo instance for ", 'red')
+        print(message + repo_path_output)
+        print_error(err)
+        sys.exit(1)
+    else:
+        return repo
+
+def _truncate_ref(ref):
+    """Return bare branch, tag, or sha"""
+    git_branch = "refs/heads/"
+    git_tag = "refs/tags/"
+    if ref.startswith(git_branch):
+        length = len(git_branch)
+    elif ref.startswith(git_tag):
+        length = len(git_tag)
+    else:
+        length = 0
+    return ref[length:]
